Rename and move `BotAPIUserKinds`
