+++ botx/bot/api/commands/system_events/chat_created.py
@@ -4,12 +4,13 @@ from uuid import UUID
 from pydantic import Field
 
 from botx.bot.api.commands.base import BotAPIBaseCommand, BotAPIChatEventSender
-from botx.bot.api.enums import BotAPICommandTypes, BotAPIUserKinds, convert_user_kind
+from botx.bot.api.enums import BotAPICommandTypes
 from botx.bot.models.commands.chat import Chat
 from botx.bot.models.commands.system_events.chat_created import (
     ChatCreatedEvent,
     ChatCreatedMember,
 )
+from botx.shared_models.api.enums import APIUserKinds, convert_user_kind
 from botx.shared_models.api_base import VerifiedPayloadBaseModel
 from botx.shared_models.chat_types import APIChatTypes, convert_chat_type_to_domain
 
@@ -23,7 +24,7 @@ class BotAPIChatMember(VerifiedPayloadBaseModel):
     is_admin: bool = Field(..., alias="admin")
     huid: UUID
     name: Optional[str]
-    user_kind: BotAPIUserKinds
+    user_kind: APIUserKinds
 
 
 class BotAPIChatCreatedData(VerifiedPayloadBaseModel):
