+++ botx/shared_models/api/enums.py
@@ -1,5 +1,11 @@
-from botx.bot.models.commands.enums import AttachmentTypes
+from botx.bot.models.commands.enums import AttachmentTypes, UserKinds
 from botx.shared_models.api_base import StrEnum
+
+
+class APIUserKinds(StrEnum):
+    USER = "user"
+    CTS_USER = "cts_user"
+    BOTX = "botx"
 
 
 class APIAttachmentTypes(StrEnum):
@@ -10,6 +16,20 @@ class APIAttachmentTypes(StrEnum):
     LOCATION = "location"
     CONTACT = "contact"
     LINK = "link"
+
+
+def convert_user_kind(user_kind: APIUserKinds) -> UserKinds:
+    user_kinds_mapping = {
+        APIUserKinds.USER: UserKinds.RTS_USER,
+        APIUserKinds.CTS_USER: UserKinds.CTS_USER,
+        APIUserKinds.BOTX: UserKinds.BOT,
+    }
+
+    converted_type = user_kinds_mapping.get(user_kind)
+    if converted_type is None:
+        raise NotImplementedError(f"Unsupported user kind: {user_kind}")
+
+    return converted_type
 
 
 def convert_attachment_type_to_domain(
